# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        minLength: 2
        type: string
      last_name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      tenant_name:
        minLength: 2
        type: string
      tenant_slug:
        minLength: 2
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - tenant_name
    - tenant_slug
    type: object
  handlers.ResetPasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - confirm_password
    - new_password
    - token
    type: object
  handlers.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      first_name:
        type: string
      language:
        type: string
      last_name:
        type: string
      phone:
        type: string
      preferences:
        $ref: '#/definitions/models.UserPreferences'
      timezone:
        type: string
    type: object
  models.Notification:
    properties:
      action_url:
        type: string
      archived_at:
        type: string
      comment:
        $ref: '#/definitions/models.TaskComment'
      comment_id:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      data:
        $ref: '#/definitions/models.NotificationData'
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      message:
        type: string
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      read_at:
        type: string
      status:
        $ref: '#/definitions/models.NotificationStatus'
      task:
        $ref: '#/definitions/models.Task'
      task_id:
        description: Related entity references
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
...
  models.Task:
    properties:
      activities:
        items:
          $ref: '#/definitions/models.TaskActivity'
        type: array
      actual_hours:
        type: number
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      attachments:
        items:
          $ref: '#/definitions/models.TaskAttachment'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      completed_at:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        description: Relationships
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.Task'
      parent_id:
        type: string
      priority:
        $ref: '#/definitions/models.TaskPriority'
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      subtasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.TaskActivity:
    properties:
      action:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        type: string
      field:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      new_value:
        type: string
      old_value:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskAttachment:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskComment:
    properties:
      content:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.TaskComment'
      parent_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TaskPriorityLow
    - TaskPriorityMedium
    - TaskPriorityHigh
    - TaskPriorityUrgent
...
  models.Tenant:
    properties:
      allow_registration:
        description: Settings fields (flattened)
        type: boolean
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      default_user_role:
        type: string
      domain:
        type: string
      email_notifications:
        description: Notification settings (flattened)
        type: boolean
      favicon_url:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      logo_url:
        type: string
      max_storage:
        description: 1GB in bytes
        type: integer
      max_tasks:
        type: integer
      max_users:
        type: integer
      name:
        type: string
      plan:
        $ref: '#/definitions/models.TenantPlan'
      primary_color:
        description: Branding settings (flattened)
        type: string
      require_email_verification:
        type: boolean
      secondary_color:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/models.TenantStatus'
      task_assignments:
        type: boolean
      task_auto_assignment:
        type: boolean
      task_completions:
        type: boolean
      task_due_dates:
        type: boolean
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      users:
        description: Relationships
        items:
          $ref: '#/definitions/models.User'
        type: array
      weekly_digest:
        type: boolean
    type: object
...
  models.User:
    properties:
      assigned_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      avatar:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      created_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      default_task_priority:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      enable_email_notifications:
        type: boolean
      enable_push_notifications:
        type: boolean
      first_name:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      language:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      show_completed_tasks:
        type: boolean
      status:
        $ref: '#/definitions/models.UserStatus'
      task_reminders:
        type: boolean
      task_view_mode:
        description: list, board, calendar
        type: string
      tasks_per_page:
        type: integer
      tenant:
        allOf:
        - $ref: '#/definitions/models.Tenant'
        description: Relationships
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      theme:
        description: User preferences (flattened)
        type: string
      timezone:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      weekly_digest:
        type: boolean
    type: object
  models.UserPreferences:
    properties:
      default_task_priority:
        type: string
      email_notifications:
        type: boolean
      push_notifications:
        type: boolean
      show_completed_tasks:
        type: boolean
      task_reminders:
        type: boolean
      task_view_mode:
        description: list, board, calendar
        type: string
      tasks_per_page:
        type: integer
      theme:
        description: light, dark, auto
        type: string
      weekly_digest:
        type: boolean
    type: object
  models.UserRole:
    enum:
    - admin
    - manager
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleManager
    - UserRoleUser
  models.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusSuspended
...
host: localhost:8080
info:
  contact:
    email: support@taskflow.com
    name: API Support
    url: http://www.taskflow.com/support
  description: A multi-tenant task management platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TaskFlow API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the current user's password
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate user
      tags:
      - auth
  /auth/logout:
    post:
      description: Invalidate user session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Get information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Update the current user's profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account and associated tenant
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user and tenant
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using reset token
      parameters:
      - description: Reset password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Reset password with token
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"